+++ changelog.md
@@ -14,6 +14,7 @@
 - support for getting out hashes of `KT256` in its canonical length (64 bytes)
 - property checking of hash functionality
 - fuzzing of hash functionality
+- made use-after-finalization impossible through the type system
 
 ### changed